Switch speed variable from double to float.
The accuracy of a double is generally not needed since the Monogame/XNA libraries only deal in single-precision `float` values.

+++ Vale/GameObjects/Actors/GameActor.cs
@@ -23,7 +23,7 @@ namespace Vale.GameObjects.Actors
         
 
 
-        public double Speed { get; set; }
+        public float Speed { get; set; }
 
         private double maxHealth;
